latest tweak to code width
for now just removed the early fixed width specification for code out
That seems to get % of screen width for the code box, so it works perfectly with maximized window.
Need to keep going to get it based on window width.  Don't know exactly how to shift it.  Maybe "@media only" , with "only" ??